Refactored code to suit the PEP8 Standards
