+++ .gitignore
@@ -9,3 +9,4 @@ migrations
 htmlcov
 node_modules
 package-lock.json
+*.yml
